generalized new item input for use with ingredients
